safari-600.7-branch/LayoutTests/media/track/track-webvtt-tc031-cue-recovery-expected.txt

Tests that a cue is recovered when a line with a "-->" is encountered (missing blank line separator).



*** Testing text track 0
EXPECTED (cues.length == '2') OK
EXPECTED (cues[0].startTime == '0') OK
EXPECTED (cues[0].endTime == '1') OK
EXPECTED (cues[0].text == 'Valid cue 1') OK
EXPECTED (cues[1].startTime == '2') OK
EXPECTED (cues[1].endTime == '3') OK
EXPECTED (cues[1].text == 'Valid cue 2') OK

*** Testing text track 1
EXPECTED (cues.length == '2') OK
EXPECTED (cues[0].startTime == '0') OK
EXPECTED (cues[0].endTime == '1') OK
EXPECTED (cues[0].text == 'Valid cue 1') OK
EXPECTED (cues[1].startTime == '2') OK
EXPECTED (cues[1].endTime == '3') OK
EXPECTED (cues[1].text == 'Valid cue 2') OK

*** Testing text track 2
EXPECTED (cues.length == '2') OK
EXPECTED (cues[0].startTime == '0') OK
EXPECTED (cues[0].endTime == '1') OK
EXPECTED (cues[0].text == 'Valid cue 1') OK
EXPECTED (cues[1].startTime == '2') OK
EXPECTED (cues[1].endTime == '3') OK
EXPECTED (cues[1].text == 'Valid cue 2') OK
END OF TEST
